This is a profanity checker which also does sentiment analysis from an audio file
This program is developed in Python in the area of Natural Language Processing. This takes input from the user in the form of live audio or an existing audio file (.wav format) and checks for any abusive words and also does the sentiment analysis of the audio (positive, neutral, negative).
This uses APIs from google and vadersentiment.